requirements.txt
# Núcleo web
Django>=5.2,<6

# Processamento de documentos
python-docx>=1.1.0
openpyxl>=3.1.2
pillow>=10.3.0           # imagens embutidas em .docx

# IA e OpenAI
openai>=1.24.0
tiktoken>=0.6.0

# Análise de dados e gráficos
pandas>=2.2.0
plotly>=5.19.0

# CLI, progresso, utilidades
tqdm>=4.66.0
python-dotenv>=1.0.1
requests>=2.32.0
httpcore>=1.0.4          # importado na view

# Opcional (DB externo)
# psycopg2-binary>=2.9.9
